Zadanie 8
A Human class is an inherited class of the Animal class.
Added interface Saleable with one method sell(Human seller, Human buyer, double price)
Implemented 'sell' method in class Animal, Phone, Car, Device.
'sell' method checks if the item to be sold is not a human, then if seller has the item to sell, then if buyer has enough cash to buy the item. When all conditions are met, the method adds 'price' to the seller cash field and subtracts 'price' from buyer cash field, then assigns the item to be sold to appropriate buyer's field and assigns null to appropriate seller's field. The 'sell' method prints appropriate information to the console.